fix container initialization
When using this Docker image with volume mounting on managed
environments (such as K8s), the directory '/usr/local/share/moin/data' is
not empty since it may contain the directory 'lost+found'.

Container initilization fails because the startup script expects the
data directory to be empty. In order to fix this bug, the script now
checks for the presence of file 'initialized', which is only created if
the data directory is fully initialiazed.